Put jore filters in one file.
